requirements.txt
aiosqlite==0.20.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.110.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.2.1
pydantic_core==2.33.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
